backend/requirements.txt
browser-use==0.1.40
anthropic>=0.40.0
langchain-anthropic>=0.3.0
openai>=1.50.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mem0ai>=0.1.29
tavily-python>=0.5.0
pymongo>=4.10.0
motor>=3.6.0
python-dotenv>=1.0.0
numpy>=2.0.0
exa-py>=1.1.7
pydantic>=2.10.0
python-multipart>=0.0.17
httpx>=0.28.0
playwright>=1.49.0
